% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_convert_data.R
\name{ck_conv_dat}
\alias{ck_conv_dat}
\title{Check point cloud suitability.}
\usage{
ck_conv_dat(data, message)
}
\arguments{
\item{data}{a data.frame or data.table.}
}
\value{
the data converted to data.table and a logical value indicating if the input data was a base R data.frame or a data.table.
}
\description{
Check point cloud suitability.
}
\keyword{internal}
