% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_convert_voxel_cloud.R
\name{ck_conv_vox}
\alias{ck_conv_vox}
\title{Check voxel cloud suitability}
\usage{
ck_conv_vox(data, message)
}
\arguments{
\item{data}{a data.frame or data.table.}

\item{message}{logical. If FALSE, messages are disabled. Default = TRUE.}
}
\value{
the data converted to data.table and a logical value indicating if the input data was a base R data.frame or a data.table.
}
\description{
Check voxel cloud suitability
}
\keyword{internal}
