% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_resolution.R
\name{guess_resolution}
\alias{guess_resolution}
\title{Guess the voxel resolution in a voxel cloud}
\usage{
guess_resolution(data, N_sample, message)
}
\arguments{
\item{data}{a data.frame or data.table containing the x, y, z, ... coordinates of a voxel cloud.}

\item{N_sample}{numeric. The number of voxels to sample. Default = 100.}

\item{message}{logical. FALSE desables the message.}
}
\value{
the guessed resolution.
}
\description{
measure the distance to the nearest neighbour for \code{N_sample} voxels and guess the resolution:
             the majority wins.
}
\keyword{internal}
